Make mechanic dropdown float
Refactored the code that did this for the callin dropdown into a mixin.

Fixes #1170
