+++ client/mechanics.less
@@ -1,5 +1,9 @@
 .bb-mechanics-dropdown {
   text-align: left;
+  .dropdown-menu {
+    position: fixed;
+    width: max-content;
+  }
 }
 .favorite-mechanic {
   font-weight: bold;
